newshub/requirements.txt
Django>=4.2
gunicorn
whitenoise
requests
transformers
newspaper3k
torch
sentry-sdk
webpush
python-dotenv
python-jose[cryptography]
python-multipart
python-dateutil
python-magic
Pillow
celery[redis]
redis
psycopg2-binary
python-decouple
django-celery-beat
django-celery-results
django-webpush
django-crispy-forms
django-crispy-bootstrap5
sentence-transformers
numpy
pandas
scikit-learn
nltk
pywebpush
python-magic-bin; sys_platform == 'win32'
